Move ParseLink and validators into runconfig.parse where they are used.
